HomeworkN3 comments fix

+++ src/HomeworkN3/strategy/CashPayment.java
@@ -1,8 +1,10 @@
 package HomeworkN3.strategy;
 
+import java.math.BigDecimal;
+
 public class CashPayment implements PaymentStrategy {
     @Override
-    public void pay(double amount) {
+    public void pay(BigDecimal amount) {
         System.out.println("Оплата наличными: " + amount + " руб.");
     }
 }
