+++ src/HomeworkN3/strategy/PayPalPayment.java
@@ -1,8 +1,10 @@
 package HomeworkN3.strategy;
 
+import java.math.BigDecimal;
+
 public class PayPalPayment implements PaymentStrategy {
     @Override
-    public void pay(double amount) {
+    public void pay(BigDecimal amount) {
         System.out.println("Оплата через PayPal: " + amount + " руб.");
     }
 }
