+++ src/HomeworkN3/strategy/PaymentStrategy.java
@@ -1,5 +1,7 @@
 package HomeworkN3.strategy;
 
+import java.math.BigDecimal;
+
 public interface PaymentStrategy {
-    void pay(double amount);
+    void pay(BigDecimal amount);
 }
